Moving EOG test for lambdas from lambda tests to EOG tests
